requirements.txt
# LLM & Embeddings
openai>=1.0.0
anthropic>=0.18.0
sentence-transformers>=2.2.0

# Azure OpenAI
azure-identity>=1.15.0

# Vector Databases
faiss-cpu>=1.7.4
chromadb>=0.4.22

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Document Processing
PyPDF2>=3.0.0
pypdf>=3.17.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
streamlit>=1.28.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
tiktoken>=0.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
